Update organizerguide.md
Hi @babaks77, 

In general, I suggest to keep the wording broad. The first event was a conference session, but new organizers might have in mind a workshop or a seminar instead, so 'event' should cover it all.

At some point any event will need to appear on a web page with a registration form, and this should go in the instructions. It could say for example 'publish the information about the event in a dedicated page of the VSF's website. The VSF can help with the creation of the forms for registration and abstract submission'.  Is this the idea?

Will events be run from the VSF (that is, yours) Zoom account? If yes it would be a useful detail to add. The point could be 'no need to worry about the online platform, the event can be run using the VSF Zoom account'. 

'Summarize your experience' makes sense only if we ask organizers to prepare some kind of report in the end, otherwise it can be left implicit that they can wrap up their thoughts in a blog post or on their website if they wish.

I would have a separate point for testing the software -- please edit to reflect what is actually being done (test the day before? Test 10 minutes before the event?)

So far speakers have been quite willing to present at virtual events, so I would remove the part that says 'approach groups of authors instead of individual authors'. We might say something about ensuring a good gender, geographical and seniority balance instead if we wish. Or we leave it up to the organizers?

I will look at the rest of the document later.

I'm still not sure if this is the best way of making changes and adding comments -- as a test I'm creating a pull request and see what happens :)